LayoutTests/editing/mac/selection/doubleclick-should-not-expand-across-lines-expected.txt

Ensure that double clicking does not expand selection across new line boundaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String(window.getSelection()) is "more"
PASS String(window.getSelection()) is "more"
PASS textarea.value.substring(textarea.selectionStart, textarea.selectionEnd) is "more"
PASS successfullyParsed is true

TEST COMPLETE
